mv `make_df` to `tests._util`
