+++ apis/python/tests/_util.py
@@ -1,6 +1,6 @@
 from contextlib import contextmanager, nullcontext
 from pathlib import Path
-from typing import Any, Tuple, Type, Union
+from typing import Any, List, Optional, Tuple, Type, Union
 
 import numpy as np
 import pandas as pd
@@ -66,6 +66,35 @@ def assert_adata_equal(ad0: AnnData, ad1: AnnData):
     assert_array_dicts_equal(ad0.varm, ad1.varm)
     assert_array_dicts_equal(ad0.obsp, ad1.obsp)
     assert_array_dicts_equal(ad0.varp, ad1.varp)
+
+
+def parse_col(col_str: str) -> Tuple[Optional[str], List[str]]:
+    """Parse a "column string" of the form ``val1,val2,...`` or ``name=val1,val2,...``."""
+    pcs = col_str.split("=")
+    if len(pcs) == 1:
+        return None, col_str.split(",")
+    elif len(pcs) == 2:
+        name, vals_str = pcs
+        vals = vals_str.split(",")
+        return name, vals
+    else:
+        raise ValueError(f"Invalid column string: {col_str}")
+
+
+def make_df(index_str: Optional[str] = None, **cols) -> pd.DataFrame:
+    """DataFrame construction helper, for tests.
+
+    - index and columns are provided as strings of the form ``name=val1,val2,...``.
+    - ``name=`` is optional for the initial (``index_str``) arg.
+    """
+    cols = dict([(col, parse_col(col_str)[1]) for col, col_str in cols.items()])
+    index = None
+    index_name = None
+    if index_str:
+        index_name, index = parse_col(index_str)
+    df = pd.DataFrame(cols, index=index)
+    df.index.name = index_name
+    return df
 
 
 HERE = Path(__file__).parent
